requirements.txt
google-generativeai==0.3.2
chromadb==0.4.22
sentence-transformers==2.2.2
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
streamlit==1.29.0
tiktoken==0.5.2